+++ backend/src/main/java/com/roc/app/competition/referee/RefereeNotEligibleException.java
@@ -1,7 +1,7 @@
 package com.roc.app.competition.referee;
 
 public class RefereeNotEligibleException extends RuntimeException {
-    public RefereeNotEligibleException(Long refereeId) {
+    public RefereeNotEligibleException(Integer refereeId) {
             super("Referee with id " + refereeId + " doesn't have proper permissions (licence type) to referee the competition");
     }
 }
